refactored folder struct
